use base instrument for kinesis and CC
